refactor(vm): split bytecode validation
